[sailfish-weather] Position provider logo consistently. JB#58767
The position of the provider logo on the page depended on the content
height of the flickable, but the content height also apparently included
the logo, causing its position to jump under certain circumstances
(removing all-but-one location from the model).

This change removes the logo from the flickable content so that it no
longer jumps position.